@@ -1,17 +1,37 @@
 name: psyclinicai
-description: Minimal sanity check
+description: "A psychology clinic management application with AI integration"
 publish_to: 'none'
 
+version: 1.0.0+1
+
 environment:
-  sdk: '>=3.3.0 <4.0.0'
+  sdk: '>=3.8.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
+  cupertino_icons: ^1.0.2
+  http: ^1.1.0
+  shared_preferences: ^2.2.2
+  crypto: ^3.0.3
+  json_annotation: ^4.9.0
+  provider: ^6.1.1
+  # PDF Export için gerekli paketler
+  pdf: ^3.10.7
+  path_provider: ^2.1.2
+  open_file: ^3.3.2
+  share_plus: ^7.2.1
+  # Image handling için
+  image: ^4.1.3
+  # QR Code için
+  qr_flutter: ^4.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_lints: ^3.0.0
+  json_serializable: ^6.7.1
+  build_runner: ^2.4.7
 
 flutter:
   uses-material-design: true
