@@ -1,15 +1,38 @@
 import 'package:flutter/material.dart';
+import 'dart:io';
+import '../../services/pdf_export_service.dart';
 import '../../utils/theme.dart';
+import '../../models/ai_response_models.dart';
 
-class PDFExportPanel extends StatelessWidget {
-  final VoidCallback onExport;
-  final bool isExporting;
+class PDFExportPanel extends StatefulWidget {
+  final String sessionNotes;
+  final SessionSummaryResponse? aiSummary;
+  final String? clientName;
+  final String? therapistName;
+  final Map<String, dynamic>? clientInfo;
+  final Map<String, dynamic>? sessionMetrics;
 
   const PDFExportPanel({
     super.key,
-    required this.onExport,
-    required this.isExporting,
+    required this.sessionNotes,
+    this.aiSummary,
+    this.clientName,
+    this.therapistName,
+    this.clientInfo,
+    this.sessionMetrics,
   });
+
+  @override
+  State<PDFExportPanel> createState() => _PDFExportPanelState();
+}
+
+class _PDFExportPanelState extends State<PDFExportPanel> {
+  final PDFExportService _pdfService = PDFExportService();
+  bool _isGeneratingPDF = false;
+  bool _isOpeningPDF = false;
+  bool _isSharingPDF = false;
+  File? _generatedPDF;
+  String? _error;
 
   @override
   Widget build(BuildContext context) {
@@ -18,24 +41,28 @@ class PDFExportPanel extends StatelessWidget {
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          // Başlık
+          // Header
           Row(
             children: [
               Icon(
                 Icons.picture_as_pdf,
-                color: AppTheme.accentColor,
+                color: AppTheme.primaryColor,
                 size: 24,
               ),
               const SizedBox(width: 8),
               Text(
                 'PDF Export',
-                style: Theme.of(context).textTheme.headlineSmall,
+                style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                  fontWeight: FontWeight.bold,
+                  color: AppTheme.primaryColor,
+                ),
               ),
             ],
           ),
+          
           const SizedBox(height: 16),
-
-          // Export seçenekleri
+          
+          // Content
           Expanded(
             child: Container(
               width: double.infinity,
@@ -47,99 +74,7 @@ class PDFExportPanel extends StatelessWidget {
                   color: Theme.of(context).dividerColor,
                 ),
               ),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  // Export formatları
-                  Text(
-                    'Export Formatları',
-                    style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                          fontWeight: FontWeight.w600,
-                        ),
-                  ),
-                  const SizedBox(height: 12),
-
-                  // Format seçenekleri
-                  _buildFormatOption(
-                    context,
-                    'Klinik Not',
-                    'Seans notu + AI özeti',
-                    Icons.medical_services,
-                    true,
-                  ),
-                  const SizedBox(height: 8),
-
-                  _buildFormatOption(
-                    context,
-                    'Danışan Raporu',
-                    'Sadece seans notu',
-                    Icons.person,
-                    false,
-                  ),
-                  const SizedBox(height: 8),
-
-                  _buildFormatOption(
-                    context,
-                    'Süpervizyon',
-                    'Detaylı analiz + öneriler',
-                    Icons.supervisor_account,
-                    false,
-                  ),
-
-                  const Spacer(),
-
-                  // Export butonu
-                  SizedBox(
-                    width: double.infinity,
-                    child: ElevatedButton.icon(
-                      onPressed: isExporting ? null : onExport,
-                      icon: isExporting
-                          ? const SizedBox(
-                              width: 16,
-                              height: 16,
-                              child: CircularProgressIndicator(strokeWidth: 2),
-                            )
-                          : const Icon(Icons.download),
-                      label: Text(
-                          isExporting ? 'PDF Oluşturuluyor...' : 'PDF Export'),
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: AppTheme.accentColor,
-                        foregroundColor: Colors.white,
-                      ),
-                    ),
-                  ),
-
-                  const SizedBox(height: 12),
-
-                  // Bilgi metni
-                  Container(
-                    padding: const EdgeInsets.all(12),
-                    decoration: BoxDecoration(
-                      color: AppTheme.accentColor.withValues(alpha: 0.1),
-                      borderRadius: BorderRadius.circular(8),
-                    ),
-                    child: Row(
-                      children: [
-                        Icon(
-                          Icons.info_outline,
-                          color: AppTheme.accentColor,
-                          size: 16,
-                        ),
-                        const SizedBox(width: 8),
-                        Expanded(
-                          child: Text(
-                            'PDF\'ler otomatik olarak kaydedilir ve paylaşılabilir',
-                            style:
-                                Theme.of(context).textTheme.bodySmall?.copyWith(
-                                      color: AppTheme.accentColor,
-                                    ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ],
-              ),
+              child: _buildContent(context),
             ),
           ),
         ],
@@ -147,62 +82,559 @@ class PDFExportPanel extends StatelessWidget {
     );
   }
 
-  Widget _buildFormatOption(
-    BuildContext context,
-    String title,
-    String description,
-    IconData icon,
-    bool isSelected,
-  ) {
-    return Container(
-      padding: const EdgeInsets.all(12),
-      decoration: BoxDecoration(
-        color: isSelected
-            ? AppTheme.accentColor.withValues(alpha: 0.1)
-            : Colors.transparent,
-        borderRadius: BorderRadius.circular(8),
-        border: Border.all(
-          color: isSelected ? AppTheme.accentColor : Colors.grey[300]!,
+  Widget _buildContent(BuildContext context) {
+    if (_isGeneratingPDF) {
+      return _buildGeneratingState(context);
+    } else if (_error != null) {
+      return _buildErrorState(context);
+    } else if (_generatedPDF != null) {
+      return _buildSuccessState(context);
+    } else {
+      return _buildInitialState(context);
+    }
+  }
+
+  Widget _buildInitialState(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text(
+          'Seans notlarınızı profesyonel PDF raporu olarak dışa aktarın',
+          style: Theme.of(context).textTheme.bodyMedium,
         ),
-      ),
-      child: Row(
-        children: [
-          Icon(
-            icon,
-            color: isSelected ? AppTheme.accentColor : Colors.grey[600],
-            size: 20,
+        
+        const SizedBox(height: 20),
+        
+        // Export Options
+        _buildExportOption(
+          context,
+          icon: Icons.description,
+          title: 'Seans Raporu',
+          description: 'Seans notları ve AI analizi ile birlikte',
+          onTap: _generateSessionReport,
+        ),
+        
+        const SizedBox(height: 12),
+        
+        _buildExportOption(
+          context,
+          icon: Icons.medical_services,
+          title: 'Tedavi Planı',
+          description: 'Tanı, hedefler ve müdahaleler',
+          onTap: _generateTreatmentPlan,
+        ),
+        
+        const SizedBox(height: 12),
+        
+        _buildExportOption(
+          context,
+          icon: Icons.trending_up,
+          title: 'İlerleme Raporu',
+          description: 'Seans geçmişi ve ilerleme metrikleri',
+          onTap: _generateProgressReport,
+        ),
+        
+        const SizedBox(height: 20),
+        
+        // Info
+        Container(
+          padding: const EdgeInsets.all(12),
+          decoration: BoxDecoration(
+            color: AppTheme.infoColor.withValues(alpha: 0.1),
+            borderRadius: BorderRadius.circular(8),
+            border: Border.all(color: AppTheme.infoColor.withValues(alpha: 0.3)),
           ),
-          const SizedBox(width: 12),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  title,
-                  style: Theme.of(context).textTheme.titleSmall?.copyWith(
-                        fontWeight: FontWeight.w600,
-                        color: isSelected ? AppTheme.accentColor : null,
-                      ),
-                ),
-                Text(
-                  description,
+          child: Row(
+            children: [
+              Icon(
+                Icons.info_outline,
+                color: AppTheme.infoColor,
+                size: 20,
+              ),
+              const SizedBox(width: 8),
+              Expanded(
+                child: Text(
+                  'PDF\'ler cihazınıza kaydedilir ve paylaşılabilir',
                   style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                        color: isSelected
-                            ? AppTheme.accentColor
-                            : Colors.grey[600],
-                      ),
+                    color: AppTheme.infoColor,
+                  ),
                 ),
-              ],
+              ),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget _buildGeneratingState(BuildContext context) {
+    return Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          CircularProgressIndicator(
+            valueColor: AlwaysStoppedAnimation<Color>(AppTheme.primaryColor),
+          ),
+          const SizedBox(height: 16),
+          Text(
+            'PDF oluşturuluyor...',
+            style: Theme.of(context).textTheme.titleMedium,
+          ),
+          const SizedBox(height: 8),
+          Text(
+            'Lütfen bekleyin',
+            style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+              color: Colors.grey[600],
             ),
           ),
-          if (isSelected)
-            Icon(
-              Icons.check_circle,
-              color: AppTheme.accentColor,
-              size: 20,
-            ),
         ],
       ),
     );
+  }
+
+  Widget _buildSuccessState(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        // Success message
+        Container(
+          padding: const EdgeInsets.all(16),
+          decoration: BoxDecoration(
+            color: AppTheme.successColor.withValues(alpha: 0.1),
+            borderRadius: BorderRadius.circular(8),
+            border: Border.all(color: AppTheme.successColor.withValues(alpha: 0.3)),
+          ),
+          child: Row(
+            children: [
+              Icon(
+                Icons.check_circle,
+                color: AppTheme.successColor,
+                size: 24,
+              ),
+              const SizedBox(width: 12),
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      'PDF başarıyla oluşturuldu!',
+                      style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                        color: AppTheme.successColor,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                    const SizedBox(height: 4),
+                    Text(
+                      _generatedPDF!.path.split('/').last,
+                      style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                        color: AppTheme.successColor,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ],
+          ),
+        ),
+        
+        const SizedBox(height: 20),
+        
+        // Action buttons
+        Row(
+          children: [
+            Expanded(
+              child: ElevatedButton.icon(
+                onPressed: _isOpeningPDF ? null : _openPDF,
+                icon: _isOpeningPDF
+                    ? SizedBox(
+                        width: 16,
+                        height: 16,
+                        child: CircularProgressIndicator(
+                          strokeWidth: 2,
+                          valueColor: AlwaysStoppedAnimation<Color>(
+                            Theme.of(context).colorScheme.onPrimary,
+                          ),
+                        ),
+                      )
+                    : const Icon(Icons.open_in_new),
+                label: Text(_isOpeningPDF ? 'Açılıyor...' : 'PDF\'i Aç'),
+                style: ElevatedButton.styleFrom(
+                  backgroundColor: AppTheme.primaryColor,
+                  foregroundColor: Colors.white,
+                  padding: const EdgeInsets.symmetric(vertical: 12),
+                ),
+              ),
+            ),
+            const SizedBox(width: 12),
+            Expanded(
+              child: ElevatedButton.icon(
+                onPressed: _isSharingPDF ? null : _sharePDF,
+                icon: _isSharingPDF
+                    ? SizedBox(
+                        width: 16,
+                        height: 16,
+                        child: CircularProgressIndicator(
+                          strokeWidth: 2,
+                          valueColor: AlwaysStoppedAnimation<Color>(
+                            Theme.of(context).colorScheme.onPrimary,
+                          ),
+                        ),
+                      )
+                    : const Icon(Icons.share),
+                label: Text(_isSharingPDF ? 'Paylaşılıyor...' : 'Paylaş'),
+                style: ElevatedButton.styleFrom(
+                  backgroundColor: AppTheme.accentColor,
+                  foregroundColor: Colors.white,
+                  padding: const EdgeInsets.symmetric(vertical: 12),
+                ),
+              ),
+            ),
+          ],
+        ),
+        
+        const SizedBox(height: 16),
+        
+        // New PDF button
+        SizedBox(
+          width: double.infinity,
+          child: OutlinedButton.icon(
+            onPressed: _resetState,
+            icon: const Icon(Icons.add),
+            label: const Text('Yeni PDF Oluştur'),
+            style: OutlinedButton.styleFrom(
+              padding: const EdgeInsets.symmetric(vertical: 12),
+              side: BorderSide(color: AppTheme.primaryColor),
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget _buildErrorState(BuildContext context) {
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        Icon(
+          Icons.error_outline,
+          color: AppTheme.errorColor,
+          size: 48,
+        ),
+        const SizedBox(height: 16),
+        Text(
+          'PDF oluşturulamadı',
+          style: Theme.of(context).textTheme.titleMedium?.copyWith(
+            color: AppTheme.errorColor,
+          ),
+        ),
+        const SizedBox(height: 8),
+        Text(
+          _error ?? 'Bilinmeyen bir hata oluştu',
+          style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+            color: Colors.grey[600],
+          ),
+          textAlign: TextAlign.center,
+        ),
+        const SizedBox(height: 20),
+        ElevatedButton.icon(
+          onPressed: _resetState,
+          icon: const Icon(Icons.refresh),
+          label: const Text('Tekrar Dene'),
+          style: ElevatedButton.styleFrom(
+            backgroundColor: AppTheme.primaryColor,
+            foregroundColor: Colors.white,
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget _buildExportOption(
+    BuildContext context, {
+    required IconData icon,
+    required String title,
+    required String description,
+    required VoidCallback onTap,
+  }) {
+    return InkWell(
+      onTap: onTap,
+      borderRadius: BorderRadius.circular(8),
+      child: Container(
+        padding: const EdgeInsets.all(16),
+        decoration: BoxDecoration(
+          color: Theme.of(context).colorScheme.surface,
+          borderRadius: BorderRadius.circular(8),
+          border: Border.all(
+            color: Theme.of(context).dividerColor,
+          ),
+        ),
+        child: Row(
+          children: [
+            Container(
+              padding: const EdgeInsets.all(8),
+              decoration: BoxDecoration(
+                color: AppTheme.primaryColor.withValues(alpha: 0.1),
+                borderRadius: BorderRadius.circular(6),
+              ),
+              child: Icon(
+                icon,
+                color: AppTheme.primaryColor,
+                size: 20,
+              ),
+            ),
+            const SizedBox(width: 12),
+            Expanded(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    title,
+                    style: Theme.of(context).textTheme.titleSmall?.copyWith(
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                  const SizedBox(height: 4),
+                  Text(
+                    description,
+                    style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                      color: Colors.grey[600],
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            Icon(
+              Icons.arrow_forward_ios,
+              color: Colors.grey[400],
+              size: 16,
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  // Generate Session Report
+  Future<void> _generateSessionReport() async {
+    setState(() {
+      _isGeneratingPDF = true;
+      _error = null;
+    });
+
+    try {
+      final pdf = await _pdfService.generateSessionReport(
+        clientName: widget.clientName ?? 'Danışan',
+        sessionDate: DateTime.now().toString().substring(0, 10),
+        sessionNotes: widget.sessionNotes,
+        therapistName: widget.therapistName ?? 'Terapist',
+        aiSummary: widget.aiSummary?.toString(),
+        clientInfo: widget.clientInfo,
+        sessionMetrics: widget.sessionMetrics,
+      );
+
+      setState(() {
+        _generatedPDF = pdf;
+        _isGeneratingPDF = false;
+      });
+
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: const Text('PDF başarıyla oluşturuldu'),
+          backgroundColor: AppTheme.successColor,
+        ),
+      );
+    } catch (e) {
+      setState(() {
+        _error = e.toString();
+        _isGeneratingPDF = false;
+      });
+
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text('PDF oluşturma hatası: ${e.toString()}'),
+          backgroundColor: AppTheme.errorColor,
+        ),
+      );
+    }
+  }
+
+  // Generate Treatment Plan
+  Future<void> _generateTreatmentPlan() async {
+    setState(() {
+      _isGeneratingPDF = true;
+      _error = null;
+    });
+
+    try {
+      final pdf = await _pdfService.generateTreatmentPlan(
+        clientName: widget.clientName ?? 'Danışan',
+        diagnosis: 'Depresyon (Örnek)',
+        goals: [
+          'Depresif belirtileri azaltmak',
+          'Sosyal ilişkileri güçlendirmek',
+          'Günlük aktivitelere katılımı artırmak',
+        ],
+        interventions: [
+          'CBT teknikleri',
+          'Davranış aktivasyonu',
+          'Sosyal beceri eğitimi',
+        ],
+        therapistName: widget.therapistName ?? 'Terapist',
+        notes: 'Haftalık seanslar ile 12 haftalık program',
+        timeline: {
+          'Hafta 1-4': 'Değerlendirme ve hedef belirleme',
+          'Hafta 5-8': 'Müdahale uygulama',
+          'Hafta 9-12': 'Değerlendirme ve sonlandırma',
+        },
+      );
+
+      setState(() {
+        _generatedPDF = pdf;
+        _isGeneratingPDF = false;
+      });
+
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: const Text('Tedavi planı PDF\'i oluşturuldu'),
+          backgroundColor: AppTheme.successColor,
+        ),
+      );
+    } catch (e) {
+      setState(() {
+        _error = e.toString();
+        _isGeneratingPDF = false;
+      });
+
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text('PDF oluşturma hatası: ${e.toString()}'),
+          backgroundColor: AppTheme.errorColor,
+        ),
+      );
+    }
+  }
+
+  // Generate Progress Report
+  Future<void> _generateProgressReport() async {
+    setState(() {
+      _isGeneratingPDF = true;
+      _error = null;
+    });
+
+    try {
+      final pdf = await _pdfService.generateProgressReport(
+        clientName: widget.clientName ?? 'Danışan',
+        sessions: [
+          {
+            'date': '2024-01-15',
+            'notes': 'İlk seans, değerlendirme yapıldı',
+          },
+          {
+            'date': '2024-01-22',
+            'notes': 'CBT teknikleri uygulandı',
+          },
+          {
+            'date': '2024-01-29',
+            'notes': 'Hedef belirleme ve planlama',
+          },
+        ],
+        progressMetrics: {
+          'Toplam Seans': '3',
+          'Depresyon Skoru': '15/27',
+          'Anksiyete Skoru': '12/21',
+          'İlerleme Oranı': '%65',
+        },
+        therapistName: widget.therapistName ?? 'Terapist',
+        summary: 'Danışan tedavi sürecinde olumlu ilerleme göstermektedir.',
+        recommendations: {
+          'Devam Et': 'Haftalık seanslara devam edilmeli',
+          'Ev Ödevleri': 'Günlük aktivite kayıtları tutulmalı',
+          'Takip': '2 hafta sonra değerlendirme yapılmalı',
+        },
+      );
+
+      setState(() {
+        _generatedPDF = pdf;
+        _isGeneratingPDF = false;
+      });
+
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: const Text('İlerleme raporu PDF\'i oluşturuldu'),
+          backgroundColor: AppTheme.successColor,
+        ),
+      );
+    } catch (e) {
+      setState(() {
+        _error = e.toString();
+        _isGeneratingPDF = false;
+      });
+
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text('PDF oluşturma hatası: ${e.toString()}'),
+          backgroundColor: AppTheme.errorColor,
+        ),
+      );
+    }
+  }
+
+  // Open PDF
+  Future<void> _openPDF() async {
+    if (_generatedPDF == null) return;
+
+    setState(() {
+      _isOpeningPDF = true;
+    });
+
+    try {
+      await _pdfService.openPDF(_generatedPDF!);
+    } catch (e) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text('PDF açılamadı: ${e.toString()}'),
+          backgroundColor: AppTheme.errorColor,
+        ),
+      );
+    } finally {
+      setState(() {
+        _isOpeningPDF = false;
+      });
+    }
+  }
+
+  // Share PDF
+  Future<void> _sharePDF() async {
+    if (_generatedPDF == null) return;
+
+    setState(() {
+      _isSharingPDF = true;
+    });
+
+    try {
+      await _pdfService.sharePDF(_generatedPDF!);
+    } catch (e) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text('PDF paylaşılamadı: ${e.toString()}'),
+          backgroundColor: AppTheme.errorColor,
+        ),
+      );
+    } finally {
+      setState(() {
+        _isSharingPDF = false;
+      });
+    }
+  }
+
+  // Reset state
+  void _resetState() {
+    setState(() {
+      _generatedPDF = null;
+      _error = null;
+      _isGeneratingPDF = false;
+      _isOpeningPDF = false;
+      _isSharingPDF = false;
+    });
   }
 }
