@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import '../../utils/constants.dart';
 import '../../utils/theme.dart';
+import '../../services/ai_service.dart';
+import '../../models/ai_response_models.dart';
 import '../../widgets/session/session_note_editor.dart';
 import '../../widgets/session/ai_summary_panel.dart';
 import '../../widgets/session/pdf_export_panel.dart';
@@ -14,9 +16,12 @@ class SessionScreen extends StatefulWidget {
 
 class _SessionScreenState extends State<SessionScreen> {
   final TextEditingController _noteController = TextEditingController();
-  String _aiSummary = '';
+  final AIService _aiService = AIService();
+  
+  SessionSummaryResponse? _aiSummary;
   bool _isGeneratingAI = false;
   bool _isExportingPDF = false;
+  String? _aiError;
 
   @override
   void dispose() {
@@ -35,30 +40,38 @@ class _SessionScreenState extends State<SessionScreen> {
       return;
     }
 
-    setState(() => _isGeneratingAI = true);
+    setState(() {
+      _isGeneratingAI = true;
+      _aiError = null;
+    });
 
     try {
-      // TODO: AI service entegrasyonu
-      await Future.delayed(const Duration(seconds: 3)); // Simülasyon
+      // API anahtarı kontrolü
+      final hasValidKey = await _aiService.hasValidApiKey();
+      if (!hasValidKey) {
+        _showApiKeyDialog();
+        return;
+      }
 
+      // AI özeti oluştur
+      final summary = await _aiService.generateSessionSummary(_noteController.text.trim());
+      
       setState(() {
-        _aiSummary = '''
-Duygu: Üzgün ve umutsuz
-Tema: Değersizlik hissi ve sosyal izolasyon
-ICD Önerisi: 6B00.0 (Depresif bozukluk)
-Risk Seviyesi: Orta
-Önerilen Müdahale: CBT + Sosyal destek grupları
-        '''
-            .trim();
+        _aiSummary = summary;
+        _aiError = null;
       });
 
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
-          content: const Text('AI özeti oluşturuldu'),
+          content: const Text('AI özeti başarıyla oluşturuldu'),
           backgroundColor: AppTheme.accentColor,
         ),
       );
     } catch (e) {
+      setState(() {
+        _aiError = e.toString();
+      });
+      
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
           content: Text('AI özeti hatası: ${e.toString()}'),
@@ -68,6 +81,60 @@ Risk Seviyesi: Orta
     } finally {
       setState(() => _isGeneratingAI = false);
     }
+  }
+
+  void _showApiKeyDialog() {
+    final apiKeyController = TextEditingController();
+    
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('OpenAI API Anahtarı Gerekli'),
+        content: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            const Text(
+              'AI özelliklerini kullanmak için OpenAI API anahtarınızı girmeniz gerekiyor.',
+            ),
+            const SizedBox(height: 16),
+            TextField(
+              controller: apiKeyController,
+              decoration: const InputDecoration(
+                labelText: 'OpenAI API Anahtarı',
+                hintText: 'sk-...',
+                border: OutlineInputBorder(),
+              ),
+              obscureText: true,
+            ),
+            const SizedBox(height: 8),
+            Text(
+              'API anahtarınız güvenli bir şekilde cihazınızda saklanacaktır.',
+              style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                color: Colors.grey[600],
+              ),
+            ),
+          ],
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('İptal'),
+          ),
+          ElevatedButton(
+            onPressed: () async {
+              if (apiKeyController.text.trim().isNotEmpty) {
+                await _aiService.saveApiKey(apiKeyController.text.trim());
+                Navigator.pop(context);
+                
+                // API anahtarı kaydedildikten sonra AI özetini oluştur
+                _generateAISummary();
+              }
+            },
+            child: const Text('Kaydet'),
+          ),
+        ],
+      ),
+    );
   }
 
   Future<void> _exportToPDF() async {
@@ -111,6 +178,25 @@ Risk Seviyesi: Orta
       appBar: AppBar(
         title: const Text('Seans Notu + AI Özet'),
         actions: [
+          // API durumu göstergesi
+          FutureBuilder<bool>(
+            future: _aiService.hasValidApiKey(),
+            builder: (context, snapshot) {
+              if (snapshot.hasData && snapshot.data == true) {
+                return IconButton(
+                  icon: const Icon(Icons.check_circle, color: Colors.green),
+                  onPressed: () => _showApiStatus(context),
+                  tooltip: 'API Bağlantısı Aktif',
+                );
+              } else {
+                return IconButton(
+                  icon: const Icon(Icons.error, color: Colors.orange),
+                  onPressed: () => _showApiKeyDialog(),
+                  tooltip: 'API Anahtarı Gerekli',
+                );
+              }
+            },
+          ),
           // Klavye kısayolları bilgisi
           IconButton(
             icon: const Icon(Icons.keyboard),
@@ -152,17 +238,32 @@ Risk Seviyesi: Orta
                     flex: 2,
                     child: AISummaryPanel(
                       summary: _aiSummary,
+                      error: _aiError,
                       onRegenerate: _generateAISummary,
                       isGenerating: _isGeneratingAI,
                     ),
                   ),
 
-                  // PDF Export Paneli
+                  // PDF Export Panel
                   Expanded(
                     flex: 1,
                     child: PDFExportPanel(
-                      onExport: _exportToPDF,
-                      isExporting: _isExportingPDF,
+                      sessionNotes: _noteController.text,
+                      aiSummary: _aiSummary,
+                      clientName: 'Ahmet Yılmaz', // Mock data
+                      therapistName: 'Dr. Ayşe Demir', // Mock data
+                      clientInfo: {
+                        'Yaş': '28',
+                        'Cinsiyet': 'Erkek',
+                        'Meslek': 'Mühendis',
+                        'İlk Seans': '2024-01-15',
+                      },
+                      sessionMetrics: {
+                        'Seans Süresi': '50 dakika',
+                        'Mood Skoru': '6/10',
+                        'Anksiyete Seviyesi': 'Orta',
+                        'Ev Ödevi Tamamlanma': '%80',
+                      },
                     ),
                   ),
                 ],
@@ -178,6 +279,54 @@ Risk Seviyesi: Orta
         label: const Text('AI Özeti'),
         backgroundColor: AppTheme.primaryColor,
         foregroundColor: Colors.white,
+      ),
+    );
+  }
+
+  void _showApiStatus(BuildContext context) {
+    final status = _aiService.getRateLimitStatus();
+    
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('API Durumu'),
+        content: Column(
+          mainAxisSize: MainAxisSize.min,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            _buildStatusRow('API Bağlantısı', 'Aktif', Colors.green),
+            _buildStatusRow('Kullanılan İstek', '${status['requestsUsed']}', Colors.blue),
+            _buildStatusRow('Kalan İstek', '${status['requestsRemaining']}', Colors.blue),
+            _buildStatusRow('Sıfırlama Süresi', '${status['timeUntilReset']} saniye', Colors.orange),
+            _buildStatusRow('Rate Limit', status['isLimited'] ? 'Aşıldı' : 'Normal', 
+                status['isLimited'] ? Colors.red : Colors.green),
+          ],
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('Kapat'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildStatusRow(String label, String value, Color color) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 4),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          Text(label),
+          Text(
+            value,
+            style: TextStyle(
+              fontWeight: FontWeight.bold,
+              color: color,
+            ),
+          ),
+        ],
       ),
     );
   }
